notification_service/fastapi_service/requirements.txt
fastapi==0.61.1
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.17.0
gunicorn==20.1.0
python-dotenv==0.21.0
python-jose==3.3.0
orjson==3.9.2
aio_pika==9.2.0
aiohttp==3.8.5
